requirements.txt
numpy~=1.22.2
Flask~=2.0.3
requests~=2.27.1
click~=8.0.3
PyYAML~=6.0
scipy~=1.8.0
sympy~=1.9
setuptools~=60.9.1